# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A complete REST API for managing multiple WhatsApp sessions using Go, Fiber, PostgreSQL, and whatsmeow library.\n\n## Authentication\nAll API endpoints (except /health and /swagger/*) require API key authentication.\nProvide your API key in the `Authorization` header.",
        "title": "zpwoot - WhatsApp Multi-Session API",
        "contact": {
            "name": "zpwoot Support",
            "url": "https://github.com/your-org/zpwoot",
            "email": "support@zpwoot.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/sessions/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new Wameow session with the provided configuration. Requires API key authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Create a new Wameow session",
                "parameters": [
                    {
                        "description": "Session creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Session created successfully",
                        "schema": {
                            "$ref": "#/definitions/CreateSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or parameters",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/sessions/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a list of all Wameow sessions with optional filtering. Requires API key authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "List all Wameow sessions",
                "parameters": [
                    {
                        "enum": [
                            "created",
                            "connecting",
                            "connected",
                            "disconnected",
                            "error",
                            "logged_out"
                        ],
                        "type": "string",
                        "example": "\"connected\"",
                        "description": "Filter by session status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"5511999999999@s.Wameow.net\"",
                        "description": "Filter by device JID",
                        "name": "deviceJid",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "example": 20,
                        "description": "Limit number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "example": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sessions retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/ListSessionsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/sessions/{sessionId}/connect": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Establishes connection with Wameow for the specified session. Will generate QR code if not paired. You can use either the session UUID or session name. Requires API key authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Connect Wameow session",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"mySession\"",
                        "description": "Session ID or Name",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Connection initiated successfully",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid session identifier",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/sessions/{sessionId}/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permanently removes a Wameow session and all associated data. This action cannot be undone. You can use either the session UUID or session name. Requires API key authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Delete a Wameow session",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"mySession\"",
                        "description": "Session ID or Name",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session deleted successfully",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid session identifier",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/sessions/{sessionId}/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves detailed information about a specific Wameow session including connection status and device info. You can use either the session UUID or session name. Requires API key authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Get session information",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"mySession\"",
                        "description": "Session ID or Name",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session info retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/SessionInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid session identifier",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateSessionRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3,
                    "example": "my-whatsapp-session"
                },
                "proxyConfig": {
                    "$ref": "#/definitions/ProxyConfig"
                }
            }
        },
        "CreateSessionResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "1b2e424c-a2a0-41a4-b992-15b7ec06b9bc"
                },
                "isConnected": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "my-whatsapp-session"
                },
                "proxyConfig": {
                    "$ref": "#/definitions/ProxyConfig"
                }
            }
        },
...
        "ListSessionsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/SessionInfoResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "SessionInfoResponse": {
            "type": "object",
            "properties": {
                "deviceInfo": {
                    "$ref": "#/definitions/DeviceInfoResponse"
                },
                "session": {
                    "$ref": "#/definitions/SessionResponse"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter your API key directly (no Bearer prefix required). Example: dev-api-key-12345",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}